Refactor: decouple rendering and go-openai stream
go-openai's CompletionStream is difficult to mock due to the way it is
written. In order to effectively test the (very simple) rendering
function, I moved go-openai interactions into an adapter and use a
channel to provide the token responses. This seems like overkill now,
but will ultimately make it easier to test a more complex UI.